requirements.txt
google-adk==1.1.1
litellm==1.72.0
google-genai==1.18.0
requests==2.32.3
python-dotenv==1.1.0
